+++ docs/changes.md
@@ -39,6 +39,7 @@ Changes in command line arguments:
 - [NXDRIVE-600](https://jira.nuxeo.com/browse/NXDRIVE-600): Use the generic JSP to acquire a token
 - [NXDRIVE-969](https://jira.nuxeo.com/browse/NXDRIVE-969): Switch to QML for UI
 - [NXDRIVE-1183](https://jira.nuxeo.com/browse/NXDRIVE-1183): Make server UI selection smarter
+- [NXDRIVE-1195](https://jira.nuxeo.com/browse/NXDRIVE-1195): Add a setting to let the user choose the icons set
 - [NXDRIVE-1231](https://jira.nuxeo.com/browse/NXDRIVE-1231): Add a message box to display fatal errors
 - [NXDRIVE-1235](https://jira.nuxeo.com/browse/NXDRIVE-1235): Disable Hebrew language as it is not translated
 - [NXDRIVE-1267](https://jira.nuxeo.com/browse/NXDRIVE-1267): Add a notification when DirectEdit starts the download of a file